fix: address review findings (ARCH-B-3 round 1)
