+++ src/mindroom/bot.py
@@ -41,7 +41,6 @@ from mindroom.matrix.presence import build_agent_status_message, set_presence_st
 from mindroom.matrix.room_cleanup import cleanup_all_orphaned_bots
 from mindroom.matrix.rooms import leave_non_dm_rooms, resolve_room_aliases
 from mindroom.matrix.sync_tokens import load_sync_token, save_sync_token
-from mindroom.matrix.thread_membership import resolve_related_event_thread_id_best_effort
 from mindroom.matrix.users import (
     AgentMatrixUser,
     create_agent_user,
@@ -734,7 +733,7 @@ class AgentBot:
         thread_id: str | None = None
         if normalized_target_event_id:
             try:
-                thread_id = await resolve_related_event_thread_id_best_effort(
+                thread_id = await self._conversation_resolver.resolve_related_event_thread_id_best_effort(
                     room_id,
                     normalized_target_event_id,
                     access=self._conversation_resolver.thread_membership_access(
