+++ src/mindroom/matrix/client.py
@@ -614,7 +614,6 @@ async def leave_room(client: nio.AsyncClient, room_id: str) -> bool:
 
 
 __all__ = [
-    "DeliveredMatrixEvent",
     "PermanentMatrixStartupError",
     "ResolvedVisibleMessage",
     "RoomThreadsPageError",
@@ -628,7 +627,6 @@ __all__ = [
     "invite_to_room",
     "join_room",
     "leave_room",
-    "matrix_client",
     "replace_visible_message",
     "send_file_message",
     "send_message_result",
