+++ src/mindroom/matrix/conversation_cache.py
@@ -36,10 +36,10 @@ from mindroom.matrix.client_thread_history import (
 )
 from mindroom.matrix.event_info import EventInfo
 from mindroom.matrix.message_content import extract_edit_body
-from mindroom.matrix.thread_bookkeeping import ThreadMutationResolver
 from mindroom.matrix.thread_membership import (
     ThreadMembershipAccess,
     fetch_event_info_for_client,
+    fetch_event_info_from_conversation_cache,
     lookup_thread_id_from_conversation_cache,
     resolve_event_thread_id,
     room_scan_thread_membership_access,
@@ -111,7 +111,14 @@ def _room_scan_membership_access_for_client(
     async def resolved_fetch_event_info(lookup_room_id: str, lookup_event_id: str) -> EventInfo | None:
         if fetch_event_info is not None:
             return await fetch_event_info(lookup_room_id, lookup_event_id)
-        return None
+        if conversation_cache is None:
+            return None
+        return await fetch_event_info_from_conversation_cache(
+            conversation_cache,
+            lookup_room_id,
+            lookup_event_id,
+            strict=True,
+        )
 
     return room_scan_thread_membership_access(
         lookup_thread_id=lookup_thread_id,
@@ -395,6 +402,8 @@ class MatrixConversationCache(ConversationCacheProtocol):
 
     def __post_init__(self) -> None:
         """Bind extracted read/write collaborators to this facade."""
+        from mindroom.matrix.thread_bookkeeping import ThreadMutationResolver  # noqa: PLC0415
+
         self._reads = ThreadReadPolicy(
             logger_getter=lambda: self.logger,
             runtime=self.runtime,
